# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API doc",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "test@example.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sYXJhdmVsdGVzdC5sb2NhbFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTcwOTk4NTc3NSwiZXhwIjoxNzA5OTg5Mzc1LCJuYmYiOjE3MDk5ODU3NzUsImp0aSI6ImJDYVZBZTFvVG51VTVzNjciLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.VYrNSCV_Fja2jDwSMXPJ1MzsjJuvQlXbTC-pkIriejo"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "3600"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/info": {
            "get": {
                "tags": [
                    "Info"
                ],
                "summary": "Get user info",
                "operationId": "5a4f4c98e7b284de49e041a4b66060ff",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Test User"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@example"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "example": "2024-03-07T18:01:48.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-03-07T18:01:48.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-03-07T18:01:48.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Logout"
                ],
                "summary": "Logout account",
                "operationId": "acf084742b0fca8c5e388f0cc2bf7683",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}